backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
pydantic-settings
passlib==1.7.4
PyJWT==2.8.0
reportlab==4.0.4
python-jose
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.104.0
python-multipart
python-dotenv
alembic